LayoutTests/imported/w3c/web-platform-tests/cookies/samesite/form-post-blank-reload.https-expected.txt

CONSOLE MESSAGE: [object MessageEvent]
Blocked access to external URL https://www1.localhost:9443/cookies/samesite/resources/puppet.html

Harness Error (TIMEOUT), message = null

PASS Reloaded same-host top-level form POSTs are strictly same-site
TIMEOUT Reloaded subdomain top-level form POSTs are strictly same-site Test timed out
NOTRUN Reloaded cross-site top-level form POSTs are not same-site
